requirements.txt
aioamqp==0.13.0
amqp==2.5.1
anyjson==0.3.3
asgiref==3.2.3
asn1crypto==0.24.0
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
bcrypt==3.1.6
billiard==3.6.1.0
bson==0.5.8
celery==4.3.0
certifi==2018.11.29
cffi==1.11.5
channels==2.3.1
channels-rabbitmq==1.1.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
Cython==0.29.21
daphne==2.3.0
dataclasses==0.6
defusedxml==0.5.0
Django==2.0.7
django-bootstrap4==0.0.7
django-cors-headers==2.4.0
django-extensions==2.2.1
django-filter==2.1.0
django-spa==0.3.5
django-webpack-loader==0.6.0
djangorestframework==3.9.1
djangorestframework-xml==1.4.0
djongo==1.3.1
graphviz==0.11.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.19
incremental==17.5.0
isodate==0.6.0
kombu==4.6.4
lxml==4.6.2
Markdown==3.0.1
more-itertools==7.2.0
msgpack==0.6.2
pamqp==2.3.0
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
pydotplus==2.0.2
PyHamcrest==1.9.0
pymongo==3.7.2
PyNaCl==1.3.0
pyparsing==2.4.2
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.14.0
sqlparse==0.2.4
sshtunnel==0.1.4
Twisted==19.10.0
txaio==18.8.1
urllib3==1.24.1
vine==1.3.0
whitenoise==5.0.1
zipp==0.6.0
zope.interface==4.7.1
